Move image request into WritingPage and refactor prop drilling
