+++ src/Components/Inspirations/InspirationImage.tsx
@@ -1,38 +1,18 @@
 import React, { useEffect, useState } from 'react';
-import { getImages } from './imageApiCalls';
 import './Inspirations.scss'
 import errorImage from './error-image.png'
 
-interface IImagesData {
-  data: IImageData[]
-}
-
-interface IImageData {
-  author: string,
-  download_url: string,
-  id: string,
-  height: number,
-  url: string,
-  width: number
-}
 
 interface Error {
   error: boolean
 }
 
 interface Props {
-  setImage: (arg0: string) => void
+  getImage: () => void
+  image: {author: string, download_url: string}
 }
 
-const InspirationImage: React.FC<Props> = ({ setImage }) => {
-  const [ inspirationImage, setInspirationImage ] = useState<IImageData>({
-    id: "",
-    author:"loading...",
-    width:5616,
-    height:3744,
-    url:"",
-    download_url:""
-  })
+const InspirationImage: React.FC<Props> = ({ getImage, image }) => {
 
   const [ errorHandle, setErrorHandle ] = useState<Error>({
     error: false
@@ -42,20 +22,6 @@ const InspirationImage: React.FC<Props> = ({ setImage }) => {
     getImage()
   }, []);
   
-  const getImage = () => {
-    getImages()
-    .then((data: IImageData[]) => {
-      const randomImage = randomIndex(data)
-      setInspirationImage(randomImage)
-      setImage(randomImage.download_url)
-    })
-    .catch(() => {
-      setErrorHandle({error: true})
-    })
-  }
-
-  const randomIndex = (data: IImageData[]) => data[Math.floor(Math.random() * data.length)]
-
   const getNewImage = () => {
     getImage()
   }
@@ -63,7 +29,7 @@ const InspirationImage: React.FC<Props> = ({ setImage }) => {
   if (errorHandle.error === false) {
     return (
       <>
-        <img className='img-inspo' src={inspirationImage.download_url} id={inspirationImage.id} alt={inspirationImage.author} />
+        <img className='img-inspo' src={image.download_url} alt={image.author} />
         <button onClick={() => getNewImage()}className='new-img-btn'>New Image</button>
       </>
     )
